pyproject.toml
[project]
name = "keywords-scrapper"
version = "0.1.0"
description = ""
authors = [
    {name = "MR11Robot",email = "taha.youssef.fares@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-oauth2-tool (>=0.0.3,<0.0.4)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "schedule (>=1.2.2,<2.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "flask (>=3.1.3,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    {include = "src"}
]

[tool.poetry.scripts]
app = "app:main"
